% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_qc.R
\name{boleCankersILowerQC}
\alias{boleCankersILowerQC}
\title{Return a list of trees with non-valid responses in the lower bole canker columns}
\usage{
boleCankersILowerQC()
}
\description{
Return a list of trees with non-valid responses in the lower bole canker columns
}
