% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_qc.R
\name{boleCankersIMiddleQC}
\alias{boleCankersIMiddleQC}
\title{Return a list of trees with non-valid responses in the middle bole canker columns}
\usage{
boleCankersIMiddleQC()
}
\description{
Return a list of trees with non-valid responses in the middle bole canker columns
}
