% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_qc.R
\name{boleCankersIUpperQC}
\alias{boleCankersIUpperQC}
\title{Return a list of trees with non-valid responses in the upper bole canker columns}
\usage{
boleCankersIUpperQC()
}
\description{
Return a list of trees with non-valid responses in the upper bole canker columns
}
