% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_qc.R
\name{branchCankersILowerQC}
\alias{branchCankersILowerQC}
\title{Return a list of trees with non-valid responses in the lower branch canker columns}
\usage{
branchCankersILowerQC()
}
\description{
Return a list of trees with non-valid responses in the lower branch canker columns
}
