% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_qc.R
\name{branchCankersIMiddleQC}
\alias{branchCankersIMiddleQC}
\title{Return a list of trees with non-valid responses in the middle branch canker columns}
\usage{
branchCankersIMiddleQC()
}
\description{
Return a list of trees with non-valid responses in the middle branch canker columns
}
