% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_qc.R
\name{branchCankersIUpperQC}
\alias{branchCankersIUpperQC}
\title{Return a list of trees with non-valid responses in the upper branch canker columns}
\usage{
branchCankersIUpperQC()
}
\description{
Return a list of trees with non-valid responses in the upper branch canker columns
}
