% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seedling_qc.R
\name{causeOfDeathQC}
\alias{causeOfDeathQC}
\title{Return a list of seedlings that have a missing or non-valid cause of death}
\usage{
causeOfDeathQC()
}
\description{
Return a list of seedlings that have a missing or non-valid cause of death
}
