% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{cleaningPine}
\alias{cleaningPine}
\title{Some basic cleaning of the pine data to be able to graph it}
\usage{
cleaningPine(pine)
}
\arguments{
\item{pine}{the data to input into the function (entire pine dataset)}
}
\description{
Some basic cleaning of the pine data to be able to graph it
}
