% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_qc.R
\name{crownKillLowerQC}
\alias{crownKillLowerQC}
\title{Return a list of trees whose lower crown kill is more than 100\% or live PIAL trees that have a null lower crown kill but are not null in both upper and middle crown kill percents}
\usage{
crownKillLowerQC()
}
\description{
Return a list of trees whose lower crown kill is more than 100\% or live PIAL trees that have a null lower crown kill but are not null in both upper and middle crown kill percents
}
