% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_qc.R
\name{crownKillMiddleQC}
\alias{crownKillMiddleQC}
\title{Return a list of trees whose middle crown kill is more than 100\% or live PIAL trees that have a null middle crown kill but are not null in both upper and lower crown kill percents}
\usage{
crownKillMiddleQC()
}
\description{
Return a list of trees whose middle crown kill is more than 100\% or live PIAL trees that have a null middle crown kill but are not null in both upper and lower crown kill percents
}
