% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_qc.R
\name{crownKillUpperQC}
\alias{crownKillUpperQC}
\title{Return a list of trees whose upper crown kill is more than 100\% or live PIAL trees that have a null upper crown kill but are not null in both middle and lower crown kill percents}
\usage{
crownKillUpperQC()
}
\description{
Return a list of trees whose upper crown kill is more than 100\% or live PIAL trees that have a null upper crown kill but are not null in both middle and lower crown kill percents
}
