% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_qc.R
\name{dbhQC}
\alias{dbhQC}
\title{Return a list of trees with a DBH greater than the median plus three standard deviations or missing}
\usage{
dbhQC()
}
\description{
Return a list of trees with a DBH greater than the median plus three standard deviations or missing
}
