% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seedling_qc.R
\name{duplicateSeedlingTagQC}
\alias{duplicateSeedlingTagQC}
\title{Return a list of seedling records with duplicate tags. Tags should be unique within the subplots in plots}
\usage{
duplicateSeedlingTagQC()
}
\description{
Return a list of seedling records with duplicate tags. Tags should be unique within the subplots in plots
}
