% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pine_qc.R
\name{eventDataQC}
\alias{eventDataQC}
\title{Returns any events with no photo, seedling, and/or tree data}
\usage{
eventDataQC()
}
\value{
table containing all of the pine event data quality flags
}
\description{
Returns any events with no photo, seedling, and/or tree data
}
