% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filterPine}
\alias{filterPine}
\title{Filter some or all of the pine dataset}
\usage{
filterPine(
  data_name = "all",
  network,
  park,
  sample_frame,
  panel,
  site_code,
  visit_year,
  flag,
  protected_status,
  data_processing_level,
  case_sensitive = FALSE,
  silent = FALSE
)
}
\arguments{
\item{data_name}{A string that is either a table name (see \code{\link[=validDataTables]{validDataTables()}}), "data" (all data), "metadata" (all metadata), "all" (all data and metadata)}

\item{network}{Character vector of four letter network code(s)}

\item{park}{Character vector of four letter park code(s)}

\item{sample_frame}{Character vector of sample frame(s)}

\item{panel}{Character vector of panel(s)}

\item{site_code}{Character vector of site code(s)}

\item{visit_year}{Chatacter vector of visit year(s)}

\item{flag}{Character vector of data quality flag(s)}

\item{protected_status}{Character vector of protected status(es)}

\item{data_processing_level}{Character vector of data processing level(s)}

\item{case_sensitive}{Should non-numeric filters be treated as case-sensitive?}

\item{silent}{Suppress informational messages?}
}
\value{

}
\description{
Use \code{\link[=validFilters]{validFilters()}} to see list of valid values for filter arguments.
}
