% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_data}
\alias{get_data}
\title{Get dataset from package global env}
\usage{
get_data(data_name = "all")
}
\arguments{
\item{data_name}{A string that is either a table name (see \code{\link[=validDataTables]{validDataTables()}}), "data" (all data), "metadata" (all metadata), "all" (all data and metadata)}
}
\value{
A nested list containing lists \code{data} and \code{metadata}.
}
\description{
Get dataset from package global env
}
