% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_w_err}
\alias{get_w_err}
\title{Get data table from package global env}
\usage{
get_w_err(data_name)
}
\arguments{
\item{data_name}{A string that is either a table name (see \code{\link[=validDataTables]{validDataTables()}}), "data" (all data), "metadata" (all metadata), "all" (all data and metadata)}
}
\value{
A tibble.
}
\description{
Calls get() with an error handler that reminds user to load data and defaults to using pkg env
}
