% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{getBasalArea}
\alias{getBasalArea}
\title{Calculates basal area for every tree}
\usage{
getBasalArea(treeData)
}
\arguments{
\item{treeData}{the data to input into the function (tree table of pine data)}
}
\value{
basal area calculated for every entry
}
\description{
Calculates basal area for every tree
}
