% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{getDensityDBH}
\alias{getDensityDBH}
\title{Calculate avg density of species in 5 cm DBH bins}
\usage{
getDensityDBH(treeData, areaSampled = 2500)
}
\arguments{
\item{treeData}{the data to input into the function (tree table of pine data)}

\item{areaSampled}{the area of a plot in meters^2, default is 2500
Input area of meters sampled as meters squared, function converts to hectares
TODO: make sure this can  be scaled up
TODO: could make height groups an input into the function}
}
\description{
Calculate avg density of species in 5 cm DBH bins
}
