% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{getDominance}
\alias{getDominance}
\title{Calculate dominance and relative dominance of species in plots}
\usage{
getDominance(treeData, plotArea = 2500, grouping = locationID)
}
\arguments{
\item{treeData}{the data to input into the function (tree table of pine data)}

\item{plotArea}{area of plots as meters squared, function converts to hectacres}

\item{grouping}{the variable you want to find the relative density of (eg. locationID or park)
NOTE the total sample area is calculated from 'plotArea' so if you change the grouping don't update 'plotArea'}
}
\description{
Calculate dominance and relative dominance of species in plots
}
