% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getFilter}
\alias{getFilter}
\title{Helper function for filterPine}
\usage{
getFilter(arg, no_arg = missing(arg), several.ok = TRUE)
}
\arguments{
\item{arg}{Filter argument}

\item{no_arg}{TRUE if arg is missing}

\item{several.ok}{Allow multiple filter options?}
}
\value{
NA if arg is missing, char vector of filters otherwise.
}
\description{
Helper function for filterPine
}
