% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{getFrequency}
\alias{getFrequency}
\title{Calculate the frequency and relative frequency of each species within a specified variable}
\usage{
getFrequency(
  treeData,
  transectArea = 500,
  plotArea = 2500,
  grouping = locationID
)
}
\arguments{
\item{treeData}{the data to input into the function (tree table of pine data)}

\item{transectArea}{the area of each transect, default is 500}

\item{plotArea}{the total area of all transects in a plot, default is 2500
NOTE: if you change the grouping don't change plotArea, its automatically calculated}

\item{grouping}{the variable you want to find the relative density of (eg. locationID or park)}
}
\description{
Calculate the frequency and relative frequency of each species within a specified variable
}
