% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{getImportanceValue}
\alias{getImportanceValue}
\title{Calculate the importance value of each species within a specified variable}
\usage{
getImportanceValue(treeData, ...)
}
\arguments{
\item{treeData}{the data to input into the function (tree table of pine data)}

\item{...}{You can pass in any params which are valid for the inside functions}
}
\description{
Calculate the importance value of each species within a specified variable
}
