% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{getRelativeDensity}
\alias{getRelativeDensity}
\title{Finds the density and relative density of species within a specified variable per hectare}
\usage{
getRelativeDensity(treeData, plotArea = 2500, grouping = locationID)
}
\arguments{
\item{treeData}{the data to input into the function (tree table of pine data)}

\item{plotArea}{the area of the plots sampled in m^2, default is 2500}

\item{grouping}{the variable you want to find the relative density of (eg. locationID or park)}
}
\description{
Finds the density and relative density of species within a specified variable per hectare
}
