% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{getTreeCount}
\alias{getTreeCount}
\title{Count the number of each species present within a specified variable}
\usage{
getTreeCount(treeData, grouping = locationID)
}
\arguments{
\item{treeData}{the data to input into the function (tree table of pine data)}

\item{grouping}{the variable you want to find the count of (eg. locationID or park)}
}
\description{
Count the number of each species present within a specified variable
}
