% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seedling_qc.R
\name{heightClassQC}
\alias{heightClassQC}
\title{Return a list of seedlings that are alive but the height class does not match domain values}
\usage{
heightClassQC()
}
\description{
Return a list of seedlings that are alive but the height class does not match domain values
}
