% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadPine}
\alias{loadPine}
\title{Load five needle pine data}
\usage{
loadPine(
  data_path,
  dictionary_dir = data_path,
  dictionary_filenames = c(tables = "data_dictionary_tables.txt", attributes =
    "data_dictionary_attributes.txt", categories = "data_dictionary_categories.txt")
)
}
\arguments{
\item{data_path}{Path to either a five needle pine database backend (.accdb) or a folder containing CSV data files exported by \code{fetchaccess::fetchFromAccess()}.}

\item{dictionary_dir}{Ignored unless \code{data_path} points to a folder of CSV files. The path to a folder containing data dictionary files exported by \code{fetchaccess::fetchFromAccess()}.}

\item{dictionary_filenames}{Names of data dictionary files. Typically you will want to leave these as the default.}
}
\value{
A nested list containing lists \code{data} and \code{metadata}.
}
\description{
Execute this function every time you load the fiveneedlepine package.
}
\examples{
\dontrun{
#' # Load from CSV and data dictionary files
LoadPine("M:/Monitoring/Pine/DataExports", "M:/Monitoring/Pine/DataExports/DataDictionary")

# Load from Access db
LoadPine("M:/Monitoring/Pine/DataExports/FNP_BackEnd.accdb")
}
}
