% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_qc.R
\name{mortalityYearQC}
\alias{mortalityYearQC}
\title{Return list of recently dead PIAL trees with a null mortality year}
\usage{
mortalityYearQC()
}
\description{
Return list of recently dead PIAL trees with a null mortality year
}
