% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_qc.R
\name{noSeedlingDataQC}
\alias{noSeedlingDataQC}
\title{Return list of events with no corresponding seedling data}
\usage{
noSeedlingDataQC()
}
\description{
Return list of events with no corresponding seedling data
}
