% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seedling_qc.R
\name{numberOfSubplotsQC}
\alias{numberOfSubplotsQC}
\title{Return a list of events that do not have 9 subplots}
\usage{
numberOfSubplotsQC()
}
\description{
Return a list of events that do not have 9 subplots
}
