% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seedling_qc.R
\name{recentlyDeadSeedlingQC}
\alias{recentlyDeadSeedlingQC}
\title{Returns a list of seedlings which are RD for multiple entries}
\usage{
recentlyDeadSeedlingQC()
}
\description{
Returns a list of seedlings which are RD for multiple entries
}
