% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_qc.R
\name{recentlyDeadTreeQC}
\alias{recentlyDeadTreeQC}
\title{Returns a list of trees which are RD for multiple entries}
\usage{
recentlyDeadTreeQC()
}
\description{
Returns a list of trees which are RD for multiple entries
}
