% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pine_qc.R
\name{seedlingDataQC}
\alias{seedlingDataQC}
\title{Creates an excel spreadsheet with any data quality flags present in the pine seedling table}
\usage{
seedlingDataQC()
}
\value{
table containing all the pinr seedling data quality flags
}
\description{
Creates an excel spreadsheet with any data quality flags present in the pine seedling table
}
