% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seedling_qc.R
\name{subplotQC}
\alias{subplotQC}
\title{Return a list of subplots that do not have a value between 1 and 9 or have a missing subplot number}
\usage{
subplotQC()
}
\description{
Return a list of subplots that do not have a value between 1 and 9 or have a missing subplot number
}
