% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_qc.R
\name{treeCauseOfDeathQC}
\alias{treeCauseOfDeathQC}
\title{Return a list of dead trees that have a missing cause of death or one that does not match domain values}
\usage{
treeCauseOfDeathQC()
}
\description{
Return a list of dead trees that have a missing cause of death or one that does not match domain values
}
