% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pine_qc.R
\name{treeDataQC}
\alias{treeDataQC}
\title{Creates an excel spreadsheet with any data quality flags present in the pine tree table}
\usage{
treeDataQC()
}
\value{
table containing all of the pine tree data quality flags
}
\description{
Creates an excel spreadsheet with any data quality flags present in the pine tree table
}
