% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_qc.R
\name{treeHeightQC}
\alias{treeHeightQC}
\title{Return list of trees with a height greater than the median plus three standard deviations or missing}
\usage{
treeHeightQC()
}
\description{
Return list of trees with a height greater than the median plus three standard deviations or missing
}
