% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_qc.R
\name{treeSubplotQC}
\alias{treeSubplotQC}
\title{Return a list of trees either with a missing subplot number or a subplot number not within 1-5}
\usage{
treeSubplotQC()
}
\description{
Return a list of trees either with a missing subplot number or a subplot number not within 1-5
}
