% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_qc.R
\name{treeVitalityQC}
\alias{treeVitalityQC}
\title{Return a list of trees with a missing vitality or a vitality that don't match a domain value}
\usage{
treeVitalityQC()
}
\description{
Return a list of trees with a missing vitality or a vitality that don't match a domain value
}
