% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validFilters}
\alias{validFilters}
\title{List valid values for filtering data}
\usage{
validFilters()
}
\value{
A named list of valid filters for this dataset
}
\description{
List valid values for filtering data
}
