% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validMetadataTables}
\alias{validMetadataTables}
\title{List valid metadata table names}
\usage{
validMetadataTables()
}
\value{
A character vector of valid table names
}
\description{
List valid metadata table names
}
