% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seedling_qc.R
\name{vitalityQC}
\alias{vitalityQC}
\title{Return a list of seedlings whose vitality is missing or does not match a domain value (L, RD, D)}
\usage{
vitalityQC()
}
\description{
Return a list of seedlings whose vitality is missing or does not match a domain value (L, RD, D)
}
