% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{writePine}
\alias{writePine}
\title{Write pine data to CSV}
\usage{
writePine(
  data_dir = here::here("data", "final"),
  dictionary_dir = here::here("data", "dictionary"),
  dictionary_filenames = c(tables = "data_dictionary_tables.txt", attributes =
    "data_dictionary_attributes.txt", categories = "data_dictionary_categories.txt"),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{data_dir}{Folder to store data csv's in}

\item{dictionary_dir}{Folder to store data dictionaries in}

\item{...}{Arguments to pass to \code{\link[=filterPine]{filterPine()}}}
}
\description{
Write pine data to CSV
}
