% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{writeQCToExcel}
\alias{writeQCToExcel}
\title{Write results of QC functions to an Excel file}
\usage{
writeQCToExcel(outputFileName = "pine_qc.xlsx")
}
\arguments{
\item{outputFileName}{Name of the output Excel file}
}
\description{
Write results of QC functions to an Excel file
}
